notebooks/raja/requirements.txt
openai
gradio
langchain
langchain-openai
langchain-google-genai
pydantic
mdutils
python-dotenv